refactor(starknet_integration_tests): moving all end_to_end_integration test functions to  a module
commit-id:b8a2fdfb
