+++ crates/starknet_integration_tests/Cargo.toml
@@ -15,6 +15,7 @@ blockifier.workspace = true
 cairo-lang-starknet-classes.workspace = true
 futures.workspace = true
 indexmap.workspace = true
+infra_utils.workspace = true
 mempool_test_utils.workspace = true
 papyrus_common.workspace = true
 papyrus_consensus.workspace = true
@@ -44,7 +45,6 @@ tracing.workspace = true
 
 [dev-dependencies]
 futures.workspace = true
-infra_utils.workspace = true
 itertools.workspace = true
 pretty_assertions.workspace = true
 rstest.workspace = true
