+++ crates/starknet_integration_tests/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod config_utils;
+pub mod end_to_end_integration;
 pub mod flow_test_setup;
 pub mod integration_test_setup;
 pub mod state_reader;
